LayoutTests/platform/mac-wk1/http/tests/referrer-policy-img/no-referrer/cross-origin-http-UpgradeMixedContent.https-expected.txt

CONSOLE MESSAGE: Blocked https://localhost:8443/referrer-policy/resources/image.py?expected= from asking for credentials because it is a cross-origin request.
Tests the behavior of no-referrer referrer policy when cross origin from HTTPS to HTTP, but the request is upgraded to HTTPS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is ""
PASS successfullyParsed is true

TEST COMPLETE
